refactor(tui): extract per-feature flows, plugin handler registry
framework.py was 3052 lines mixing framework primitives (drawing,
runners, input) with per-feature flows (ESP32 hub, process manager,
ADSB helpers) and a hand-maintained NATIVE_TOOLS dict. Adding a
feature meant editing framework.py in three places, and a broken
feature crashed the whole TUI because every handler loaded eagerly.

Each feature module now owns its handlers via a HANDLERS dict at
module scope. framework.py walks FEATURE_MODULES, imports each, and
merges their HANDLERS via _load_handlers(). A module that fails to
import is logged to ~/crash.log and skipped — its menu items become
silent no-ops at dispatch (visual hide is a follow-up commit).

Moved out of framework.py into their own modules:
- run_process_manager → tui/processes.py (~110 lines)
- All ESP32 hub flows → tui/esp32_hub.py (~620 lines)
  (run_esp32_hub, run_esp32_flash_picker, _confirm_flash,
   _run_threaded_flash, _esp32_install_watchdogs,
   _pick_watchdogs_variant, run_esp32_force_*, run_esp32_usb_reset,
   run_esp32_redetect, run_esp32_fw_cache_clear, run_esp32_backup,
   _esp32_menu_for, _ESP32_*_ITEMS)
- _adsb_layers_menu_entry, _adsb_fetch_hires_entry → tui/adsb_menu.py
  (~80 lines, with C_OK/C_CRIT bare-name → tui.C_OK/tui.C_CRIT bug
   fix — the originals would NameError if invoked)

Deleted from framework.py:
- NATIVE_TOOLS dict + _get_native_tools() (replaced by FEATURE_MODULES
  + _load_handlers())
- _Firmware_MP/MRD/MC trivial wrappers (inlined in esp32_hub)
- _run_mimiclaw dynamic-dispatch wrapper (mimiclaw.py exports HANDLERS
  directly now)

Existing modules grew HANDLERS exports: config_ui, tools, games,
monitor, files, network, services, radio, adsb, adsb_home_picker,
adsb_basemap_info, meshtastic_map, marauder, mimiclaw, telegram,
watchdogs.

Updated tests for the new architecture:
- test_native_tools.test_no_orphan_modules — checks FEATURE_MODULES +
  HELPER_MODULES allowlist instead of grepping framework source
- test_telegram.test_telegram_key_resolves_to_run_telegram_at_runtime
  — uses _load_handlers() instead of NATIVE_TOOLS
- test_tui_integrity.TestNativeToolImports → TestFeatureModuleImports
  — tests every FEATURE_MODULES entry imports cleanly and exports a
  non-empty HANDLERS dict
- extract_native_tool_keys → uses live _load_handlers()
- test_all_menu_refs_have_handlers — skips _gui:/_url: prefixes

Net: framework.py 3052 → 2232 (-820 lines, -27%). Same 5 pre-existing
test failures at parity (no new regressions). 1031 tests pass.
